+++ src/components/SearchForm.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import './SearchForm.css';
+import getMovies from '../utils/TheMovieDb';
 
 const escapeKeyCode = 27;
 
@@ -48,13 +49,6 @@ class SearchForm extends React.Component {
 			</div>
 		)
 	};
-}
-
-const getMovies = async () => {
-	let res = await fetch('https://api.themoviedb.org/3/movie/popular?api_key=' + process.env.REACT_APP_THEMOVIEDB_API_KEY);
-	res = await res.json();
-	const movies = res.results;
-	return movies;
 }
 
 const mapStateToProps = (state) => ({
